修改bug 还剩下登陆注册的验证和首页图标

+++ src/page/goodsManage/myGoodsList.vue
@@ -386,14 +386,21 @@ export default {
     },
     // 查看货品
     deleteCommodity(id) {
-      $http.deleteProducts(id)
-        .then((res) => {
-          if (res.status) return;
-          this.$message({
-            message: '操作成功',
-            type: 'success',
-          });
-          this.getList();
+      this.$confirm('此操作将永久删除，是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning',
+      })
+        .then(() => {
+          $http.deleteProducts(id)
+            .then((res) => {
+              if (res.status) return;
+              this.$message({
+                message: '操作成功',
+                type: 'success',
+              });
+              this.getList();
+            });
         });
     },
     getList() {
