+++ src/page/setting/supplierManagement/supplierManagement.vue
@@ -1,93 +1,85 @@
 <template>
-  <div class="addressManagement">
-    <div :class="$style.addressManagement">
-      <div :class="$style.am_main">
-        <!-- 标签页 -->
-        <el-row>
-          <!-- 点击按键 -->
-          <div :class="$style.am_operation_btn">
-              <span @click="info_add_btn">
-                <i class="iconfont">&#xe618;</i>
-                {{'添加供应商'}}
-              </span>
+          <div  class="addressManagement">
+                <div :class="$style.addressManagement">
+                      <div :class="$style.am_main">
+                        <!-- 标签页 -->
+                        <el-row>
+                          <!-- 点击按键 -->
+                          <div :class="$style.am_operation_btn">
+                              <span @click="info_add_btn">
+                                <i class="iconfont">&#xe618;</i>
+                                {{'添加供应商'}}
+                              </span>
+                          </div>
+                        <!-- 对应的标签页内容 -->
+                          <el-table   :class="$style.table_main"
+                                      :data="info_data"
+                                      border>
+                                      <el-table-column  label="#"
+                                                        header-align="center"
+                                                        align="center"
+                                                        type="index"
+                                                        width="80">
+                                      </el-table-column>
+                                      <el-table-column  label="供应商中文名"
+                                                        header-align="center"
+                                                        align="center"
+                                                        prop="name_cn">
+                                      </el-table-column>
+                                      <el-table-column  label="供应商英文名"
+                                                        header-align="center"
+                                                        align="center"
+                                                        prop="name_en">
+                                      </el-table-column>
+                                      <el-table-column  label="操作"
+                                                        header-align="center"
+                                                        width="240">
+                                                        <template   slot-scope="scope">
+                                                                    <el-button  size="mini"
+                                                                                @click="edit(scope.row)">
+                                                                                编辑
+                                                                    </el-button>
+                                                                    <el-button  size="mini"
+                                                                                type="danger"
+                                                                                @click="delete_data(scope.row)">
+                                                                                删除
+                                                                    </el-button>
+                                                        </template>
+                                      </el-table-column>
+                          </el-table>
+                          <el-pagination  :class="$style.pagination"
+                                          v-show="+total"
+                                          @current-change="handleCurrentChange"
+                                          :current-page="currentPage"
+                                          layout="total, prev, pager, next, jumper"
+                                          :total="+total">
+                          </el-pagination>
+                        </el-row>
+                      </div>
+                </div>
+                <!-- 添加供应商 -->
+                <el-dialog  :title="this.id ? '编辑供应商' : '添加供应商 '"
+                            :visible.sync="dialogVisible"
+                            width="30%">
+                            <el-form  :rules="rules"
+                                      ref="form"
+                                      :model="distributor"
+                                      label-width="140px">
+                                      <el-form-item  label="供应商中文名:"
+                                                     prop="name_cn">
+                                                     <el-input v-model="distributor.name_cn"></el-input>
+                                      </el-form-item>
+                                      <el-form-item  label="供应商英文名:"
+                                                     prop="name_en">
+                                                      <el-input v-model="distributor.name_en"></el-input>
+                                      </el-form-item>
+                            </el-form>
+                            <span   slot="footer" class="dialog-footer">
+                                    <el-button @click="dialogVisible = false">取 消</el-button>
+                                    <el-button type="primary" @click="submit_form">确 定</el-button>
+                            </span>
+                </el-dialog>
           </div>
-        <!-- 对应的标签页内容 -->
-          <el-table
-            :class="$style.table_main"
-            :data="info_data"
-            border
-          >
-            <el-table-column
-              header-align="center"
-              align="center"
-              type="index"
-              width="80"
-              label="#">
-            </el-table-column>
-            <el-table-column
-              header-align="center"
-              align="center"
-              prop="name_cn"
-              label="供应商中文名">
-            </el-table-column>
-            <el-table-column
-              header-align="center"
-              align="center"
-              prop="name_en"
-              label="供应商英文名">
-            </el-table-column>
-            <el-table-column
-              header-align="center"
-              width="240"
-              label="操作">
-              <template slot-scope="scope">
-                <el-button size="mini"
-                            @click="edit(scope.row)">
-                            编辑
-                </el-button>
-                <el-button size="mini"
-                           type="danger"
-                            @click="delete_data(scope.row)">
-                            删除
-                </el-button>
-              </template>
-            </el-table-column>
-          </el-table>
-          <el-pagination  :class="$style.pagination"
-                          v-show="+total"
-                          @current-change="handleCurrentChange"
-                          :current-page="currentPage"
-                          layout="total, prev, pager, next, jumper"
-                          :total="+total">
-          </el-pagination>
-        </el-row>
-      </div>
-    </div>
-    <!-- 添加供应商 -->
-    <el-dialog
-      :title="this.id ? '编辑供应商' : '添加供应商 '"
-      :visible.sync="dialogVisible"
-      width="30%">
-      <el-form
-        :rules="rules"
-        ref="form"
-        :model="distributor"
-        label-width="140px">
-        <el-form-item label="供应商中文名:"
-                      prop="name_cn">
-          <el-input v-model="distributor.name_cn"></el-input>
-        </el-form-item>
-        <el-form-item label="供应商英文名:"
-                      prop="name_en">
-          <el-input v-model="distributor.name_en"></el-input>
-        </el-form-item>
-      </el-form>
-      <span slot="footer" class="dialog-footer">
-        <el-button @click="dialogVisible = false">取 消</el-button>
-        <el-button type="primary" @click="submit_form">确 定</el-button>
-      </span>
-    </el-dialog>
-  </div>
 </template>
 <script>
 import $http from '@/api';
@@ -134,32 +126,25 @@ export default {
           // console.log(re.data, 'data');
           this.info_data = re.data.data;
           this.total = re.data.total;
-        })
-        .catch(() => {});
+        });
     },
     submit_form() {
       this.$refs.form.validate((validate) => {
         if (validate) {
-          // 编辑信息
-          if (+this.id) {
+          if (+this.id) { // 编辑信息
             $http.editDistributor(this.id, this.distributor)
               .then((re) => {
                 if (re.status) return;
                 this.id = '';
                 this.dialogVisible = false;
-                // 更新数据
-                this.handleCurrentChange(this.current_page);
-              })
-              .catch(() => {});
-          } else {
-            // 添加信息
+                this.handleCurrentChange(this.current_page); // 更新数据
+              });
+          } else { // 添加信息
             $http.addDistributor(this.distributor)
               .then((re) => {
                 if (re.status) return;
-                // 更新数据
-                this.get_distributor_data();
-              })
-              .catch(() => {});
+                this.handleCurrentChange(this.current_page); // 更新数据
+              });
           }
           this.dialogVisible = false;
         }
@@ -176,8 +161,7 @@ export default {
           this.info_data = res.data.data;
           // this.total = res.data.total;
           // this.current_page = res.data.current_page;
-        })
-        .catch(() => {});
+        });
     }, // 分页查询
     edit(info) {
       this.dialogVisible = true;
@@ -186,13 +170,23 @@ export default {
       this.distributor.name_en = info.name_en;
     },
     delete_data(info) {
-      $http.deleteDistributor(info.id)
+      this.$confirm('此操作将永久删除，是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning',
+      })
         .then(() => {
-          if (!status) {
-            this.get_distributor_data();
-          }
-        })
-        .catch(() => {});
+          $http.deleteDistributor(info.id)
+            .then(() => {
+              if (!status) {
+                this.$message({
+                  type: 'success',
+                  message: '操作成功!',
+                });
+                this.get_distributor_data();
+              }
+            });
+        });
     },
   },
   created() {
